services/perception/requirements.txt
ultralytics==8.1.14
opencv-python-headless==4.9.0.80
paho-mqtt==1.6.1
python-dotenv==1.0.0
